Refactor Playlist Panel and Grid with routing
